move logic to own service
